+++ libs/docs/core/message-strip/message-strip-docs.module.ts
@@ -7,6 +7,7 @@ import { MessageStripExampleComponent } from './examples/message-strip-example.c
 import { MessageStripNoIconExampleComponent } from './examples/message-strip-noicon-example.component';
 import { MessageStripWidthExampleComponent } from './examples/message-strip-width-example.component';
 import { MessageStripHeaderComponent } from './message-strip-header/message-strip-header.component';
+import { MessageStripAlertExampleComponent } from './examples/message-strip-alert-example.component';
 
 const routes: Routes = [
     {
@@ -27,7 +28,8 @@ const routes: Routes = [
         MessageStripDocsComponent,
         MessageStripExampleComponent,
         MessageStripNoIconExampleComponent,
-        MessageStripWidthExampleComponent
+        MessageStripWidthExampleComponent,
+        MessageStripAlertExampleComponent
     ],
     providers: [currentComponentProvider('message-strip')]
 })